LayoutTests/inspector/elements/delete-from-document-expected.txt

Tests that removing child from the document is handled properly in the elements panel.

      <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
- <html>
    - <head>
      </head>
    - <body onload="runTest()">
      </body>
  </html>
- <html>
    - <head>
      </head>
    - <body onload="runTest()">
      </body>
  </html>
